+++ src/Exception/KeyNotExistsException.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of php-tools.
  *
@@ -17,7 +18,7 @@ namespace Tools\Exception;
 use Exception;
 
 /**
- * "Array key does not exist" exception
+ * "Array key does not exist" exception.
  */
 class KeyNotExistsException extends Exception
 {