+++ src/Exception/NotWritableException.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * This file is part of php-tools.
  *
@@ -17,7 +18,7 @@ namespace Tools\Exception;
 use Tools\FileException;
 
 /**
- * "File or directory is not writable" exception
+ * "File or directory is not writable" exception.
  */
 class NotWritableException extends FileException
 {