+++ src/TestSuite/TestCase.php
@@ -18,7 +18,7 @@ use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 use Tools\ReflectionTrait;
 
 /**
- * TestCase class
+ * TestCase class.
  */
 abstract class TestCase extends PHPUnitTestCase
 {
@@ -26,7 +26,9 @@ abstract class TestCase extends PHPUnitTestCase
     use TestTrait;
 
     /**
-     * Teardown any static object changes and restore them
+     * Teardown any static object changes and restore them.
+     *
+     * It empties the temporary files directory.
      * @return void
      */
     public function tearDown()
@@ -36,5 +38,19 @@ abstract class TestCase extends PHPUnitTestCase
         if (add_slash_term(TMP) !== add_slash_term(sys_get_temp_dir())) {
             unlink_recursive(TMP);
         }
+    }
+
+    /**
+     * Sets up an expectation for an exception to be raised by the code under test.
+     *
+     * This provides backward compatibility for versions of `phpunit` lower than 8.5.
+     * @param string $regularExpression Expected regular expression for the exception message
+     * @return void
+     * @todo To be removed in a future release
+     */
+    public function expectExceptionMessageMatches(string $regularExpression): void
+    {
+        $methodToCall = method_exists(PHPUnitTestCase::class, 'expectExceptionMessageMatches') ? [parent::class, 'expectExceptionMessageMatches'] : [$this, 'expectExceptionMessageRegExp'];
+        call_user_func($methodToCall, $regularExpression);
     }
 }